+++ src/App.js
@@ -1,6 +1,4 @@
 import './App.css';
-import CreateAccountPage from './components/CreateAccountPage.js'
-import LoginPage from './components/LoginPage';
 import SideNav from './components/SideNav';
 import { Outlet } from 'react-router';
